Add a favicon to the web UI
This change adds a rel="shortcut icon" to all the .xhtml pages. The
icon.png file is derived by me from the SVG version of the image kpreid2
attached to ticket #1185, available at [0], licensed by the author,
Kevin Reid, under the CC BY 3.0 Unported license in #1185:7. My
icon-sized derivative is available at [1] and [2] and is licensed by me
under the CC BY 3.0 license in #1185:16. Note that, as far as is
possible on account of its being a derivative work, the The Tahoe-LAFS
Software Foundation also holds the right to relicense the logo under any
license it chooses as it is a contribution to Tahoe-LAFS. IANAL, even if
I start to sound like one after a while.

[0] http://switchb.org/kpreid/2009/tahoe/logo-2.svg
[1] http://marcus.wanners.net/tahoe-icon.svg
[2] http://marcus.wanners.net/tahoe-icon.png